+++ stdlib/Random/src/misc.jl
@@ -52,7 +52,7 @@ number generator, see [Random Numbers](@ref).
 
 # Examples
 ```jldoctest
-julia> srand(0); randstring()
+julia> Random.seed!(0); randstring()
 "0IPrGg0J"
 
 julia> randstring(MersenneTwister(0), 'a':'z', 6)