+++ stdlib/SparseArrays/test/sparse.jl
@@ -7,7 +7,7 @@ using SparseArrays
 using LinearAlgebra
 using Base.Printf: @printf
 using Random
-using Test: guardsrand
+using Test: guardseed
 
 @testset "issparse" begin
     @test issparse(sparse(fill(1,5,5)))
@@ -1442,7 +1442,7 @@ end
 end
 
 @testset "droptol" begin
-    local A = guardsrand(1234321) do
+    local A = guardseed(1234321) do
         triu(sprand(10, 10, 0.2))
     end
     @test SparseArrays.droptol!(A, 0.01).colptr == [1,1,1,2,2,3,4,6,6,7,9]
@@ -1742,13 +1742,13 @@ end
 end
 
 @testset "sparse matrix opnormestinv" begin
-    srand(1234)
+    Random.seed!(1234)
     Ac = sprandn(20,20,.5) + im* sprandn(20,20,.5)
     Aci = ceil.(Int64, 100*sprand(20,20,.5)) + im*ceil.(Int64, sprand(20,20,.5))
     Ar = sprandn(20,20,.5)
     Ari = ceil.(Int64, 100*Ar)
     if Base.USE_GPL_LIBS
-        # NOTE: opnormestinv is probabilistic, so requires a fixed seed (set above in srand(1234))
+        # NOTE: opnormestinv is probabilistic, so requires a fixed seed (set above in Random.seed!(1234))
         @test SparseArrays.opnormestinv(Ac,3) ≈ opnorm(inv(Array(Ac)),1) atol=1e-4
         @test SparseArrays.opnormestinv(Aci,3) ≈ opnorm(inv(Array(Aci)),1) atol=1e-4
         @test SparseArrays.opnormestinv(Ar) ≈ opnorm(inv(Array(Ar)),1) atol=1e-4
@@ -1795,7 +1795,7 @@ end
 end
 
 @testset "factorization" begin
-    srand(123)
+    Random.seed!(123)
     local A
     A = sparse(Diagonal(rand(5))) + sprandn(5, 5, 0.2) + im*sprandn(5, 5, 0.2)
     A = A + copy(A')
@@ -2059,7 +2059,7 @@ end
 
 @testset "reverse search direction if step < 0 #21986" begin
     local A, B
-    A = guardsrand(1234) do
+    A = guardseed(1234) do
         sprand(5, 5, 1/5)
     end
     A = max.(A, copy(A'))